requirements.txt
fastapi>=0.115.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
email-validator>=2.0.0
